+++ ext.php
@@ -12,16 +12,13 @@ namespace phpbbstudio\goa;
 
 /**
  * phpBB Studio - GitHub OAuth2 light Extension base
- *
- * It is recommended to remove this file from
- * an extension if it is not going to be used.
  */
 class ext extends \phpbb\extension\base
 {
 	/**
 	 * Check whether the extension can be enabled.
 	 * Provides meaningful(s) error message(s) and the back-link on failure.
-	 * CLI and 3.1/3.2 compatible (we do not use the $lang object here on purpose)
+	 * CLI compatible.
 	 *
 	 * @return bool
 	 */
@@ -33,9 +30,9 @@ class ext extends \phpbb\extension\base
 		$user->add_lang_ext('phpbbstudio/goa', 'ext_require');
 		$lang = $user->lang;
 
-		if (!(phpbb_version_compare(PHPBB_VERSION, '3.2.7', '>=') && phpbb_version_compare(PHPBB_VERSION, '3.3.0@dev', '<')))
+		if (!(phpbb_version_compare(PHPBB_VERSION, '3.2.7', '>=') && phpbb_version_compare(PHPBB_VERSION, '4.0.0@dev', '<')))
 		{
-			$lang['EXTENSION_NOT_ENABLEABLE'] .= '<br>' . $user->lang('ERROR_PHPBB_VERSION', '3.2.7', '3.3.0@dev');
+			$lang['EXTENSION_NOT_ENABLEABLE'] .= '<br>' . $user->lang('ERROR_PHPBB_VERSION', '3.2.7', '4.0.0@dev');
 			$is_enableable = false;
 		}