+++ language/en/ext_require.php
@@ -15,16 +15,15 @@ if (!defined('IN_PHPBB'))
 
 if (empty($lang) || !is_array($lang))
 {
-	$lang = array();
+	$lang = [];
 }
 
 /**
 * Some characters you may want to copy&paste: ’ » “ ” …
 */
-
-$lang = array_merge($lang, array(
+$lang = array_merge($lang, [
 	'ERROR_PHPBB_VERSION'	=> 'Minimum phpBB version required is %1$s but less than %2$s',
 	'ERROR_PHP_CURL'		=> 'PHP’s extension “CURL” must be loaded and active for this extension to work!',
 	'ERROR_PHP_INTL'		=> 'PHP’s extension “INTL” must be loaded and active for this extension to work!',
 	'ERROR_SSL_STREAM'		=> 'This site is not SSL capable.',
-));
+]);