+++ docs/CHANGELOG.md
@@ -1,8 +1,17 @@
-v1.0.0-beta
-- first release
+# phpBB Studio - GitHub OAuth2 light
+
+v1.0.3 26-08-2019
+- Use our custom GitHub class written from scratch
+- Use of Composer autoloader
+- Remove GitHub OAuth's data from tables upon Delete Data action
+- OAuth2 - Prevent CSRF and Clickjacking
+- Code housekeeping
+
+v1.0.2-beta
+- fix configurations
 
 v1.0.1-beta
 - fix class names core bug (work-around)
 
-v1.0.2-beta
-- fix configurations
+v1.0.0-beta
+- first release